Handled API Response and Error Handling
Imported cookies-parser to access client browser and CORS for allowing requests from a particular source. Currently, it is set up to receive requests from all. Used the wrapper for async handling so that we don't have to specify the error and catch block repeatedly. This also helps for middleware (req, res, next) in util. ApiError and ApiResponse have overridden the Error and Response classes.

Day-01/21-05-2024/Monday