Added logout controller and JWT verification
First, the JWT is verified from the cookies or authorization header (bearer token) to check if the token is valid or not. VerifyJwt acts as middleware to handle this verification process. In this, we also set req.user, which helps retrieve the user so that we can log out the logged-in user. We then set the refresh token to undefined and clear the cookies, effectively logging out the user who does not have a token.

Day-04/26-05-2024/Sunday